+++ internal/mdathome/structs.go
@@ -35,6 +35,7 @@ type ClientSettings struct {
 
 	// Security
 	AllowVisitorRefresh    bool `json:"allow_visitor_refresh"`
+	RejectInvalidHostname  bool `json:"reject_invalid_hostname"`
 	RejectInvalidSNI       bool `json:"reject_invalid_sni"`
 	RejectInvalidTokens    bool `json:"reject_invalid_tokens"`
 	SendServerHeader       bool `json:"send_server_header"`